Move @Indexed annotation to high-level annotation package
