+++ src/test/java/org/springframework/data/aerospike/sample/AutoIndexedDocument.java
@@ -2,7 +2,7 @@ package org.springframework.data.aerospike.sample;
 
 import com.aerospike.client.query.IndexType;
 import lombok.Value;
-import org.springframework.data.aerospike.index.Indexed;
+import org.springframework.data.aerospike.annotation.Indexed;
 import org.springframework.data.aerospike.mapping.Document;
 import org.springframework.data.aerospike.mapping.Field;
 import org.springframework.data.annotation.Id;
